backend/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-cors-headers==3.13.0
django-filter==21.1
django-import-export==2.8.0
django-simple-history==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
sentry-sdk==1.5.12
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
